+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/oddion/Oddion.java
@@ -88,7 +88,6 @@ public class Oddion extends PathfinderMob implements GeoEntity, VariantHolder<Od
     public Oddion(EntityType<? extends PathfinderMob> entityType, Level level) {
         super(entityType, level);
         this.onionTime = 6000;
-        setPersistenceRequired();
     }
 
     public Oddion(Level level) {
@@ -176,6 +175,12 @@ public class Oddion extends PathfinderMob implements GeoEntity, VariantHolder<Od
 //
 //        super.tick();
 //    }
+
+    @Override
+    public void tick() {
+        super.tick();
+        refreshDimensions();
+    }
 
     @Override
     public void aiStep() {
@@ -334,8 +339,14 @@ public class Oddion extends PathfinderMob implements GeoEntity, VariantHolder<Od
     }
 
     @Override
-    public void setVariant(Variant variant) {
+    public void setVariant(@NotNull Variant variant) {
         this.entityData.set(DATA_VARIANT, variant.getId());
+    }
+
+    @Override
+    public @NotNull EntityDimensions getDimensions(@NotNull Pose pose) {
+        if (isGrounded()) return this.getType().getDimensions().scale(1F, 0.25F);
+        return super.getDimensions(pose);
     }
 
     @Override
@@ -375,7 +386,7 @@ public class Oddion extends PathfinderMob implements GeoEntity, VariantHolder<Od
             return BY_ID.apply(id);
         }
 
-        private static Oddion.Variant getSpawnVariant(RandomSource random) {
+        private static Oddion.Variant getSpawnVariant(@NotNull RandomSource random) {
             int i = random.nextInt(100);
             return i < 47 ? STANDARD : i < 94 ? PINK : ALBINO;
         }